pyproject.toml
[project]
name = "rootsight"
version = "0.1.0"
description = "Zero-infrastructure AI incident intelligence engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.22",
    "httpx>=0.27.0",
    "google-generativeai>=0.8.0",
    "groq>=0.11.0",
    "faiss-cpu>=1.8.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.26",
]

[tool.setuptools.packages.find]
include = ["src*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["pytest-cache-files-*", ".pytest_cache", ".venv", ".next", "node_modules"]
addopts = "-p no:cacheprovider"
pythonpath = ["."]
